[TASK] Fetch and merge exception codes in working directory
New fetched TYPO3 exception code files get saved in the working
directory which can be different from the current files directory.
If it is different, then code files of both - files and working -
directory get considered for the final merged exception codes file
and that file gets saved to and read from the working directory.

This feature helps defining an exception codes directory outside
of the app folder as that folder gets completely overwritten with each
app deployment.